[None][test] consolidate CBTS unit tests in CPU stage
